Get rid of some compiler warnings
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@781 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseMerge/BaseView.cpp
@@ -1227,7 +1227,7 @@ BOOL CBaseView::OnToolTipNotify(UINT id, NMHDR *pNMHDR, LRESULT *pResult)
 	TOOLTIPTEXTA* pTTTA = (TOOLTIPTEXTA*)pNMHDR;
 	TOOLTIPTEXTW* pTTTW = (TOOLTIPTEXTW*)pNMHDR;
 	CString strTipText;
-	UINT nID = pNMHDR->idFrom;
+	UINT nID = (UINT)pNMHDR->idFrom;
 	if (pNMHDR->code == TTN_NEEDTEXTA && (pTTTA->uFlags & TTF_IDISHWND) ||
 		pNMHDR->code == TTN_NEEDTEXTW && (pTTTW->uFlags & TTF_IDISHWND))
 	{
